change LX PM TO INT

+++ src/main/java/com/mycompany/climate/model/real/RealParamClimate.java
@@ -24,7 +24,7 @@ public class RealParamClimate {
     private double waterTankLevelHumidityAirReal;
     private double waterTankLevelHumiditySoilReal;
     private int carbonDioxideReal;
-    private int carbonDioxideTankLevelReal;
+    private double carbonDioxideTankLevelReal;
     private int whiteLightReal;
 
     private double temperatureAirReal;
